Delete stripe_checkout.js
An older version of this file was sending me a 403 error when hitting the 'Pay Now' button on the cart page. The Stripe checkout page reads the following: "Note: Checkout must be loaded from https://checkout.stripe.com/checkout.js. Using a local copy of Checkout is unsupported, and may result in user-visible errors." https://stripe.com/docs/checkout. Removing this and using checkout.js in the main.html resolves the issue.